Fix race in video capture enable / disable.
In case we disable video capture and instantly enable video capture,
the PostTask already was invoked in setStateUpdated lambda, so
the later setStateUpdated(nullptr) didn't cancel the sent event.

After that we called instantly setOutgoingVideoState(Active),
and later after that the event was delivered,
calling setOutgoingVideoState(Inactive).

Now we use a separate explicit guard for the event, so that it
won't be delivered after setSTateUpdated(nullptr) call.